Publicly expose the NeighborList to help implementing new systems
